backend/requirements.txt
aiosqlite==0.17.0
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.4
asyncpg==0.23.0
black==21.5b2
click==8.0.1
databases[postgresql]==0.4.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.2.0
idna==3.2
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
ormar==0.10.10
pathspec==0.8.1
psycopg2-binary==2.8.6
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
